Rename SDK
