+++ android-sdk/.idea/.name
@@ -1,1 +1,1 @@
-SDK TQC Solution
+Android Mobile SDK
