+++ android-sdk/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">SDK TQC Solution</string>
+    <string name="app_name">Android Mobile SDK</string>
 </resources>
